添加 计算当前交易日期的前一个交易日的日K线数据 函数

+++ trade-biz-dal-tradecore/src/main/java/com/trade/biz/dal/tradecore/DayKLineDao.java
@@ -20,7 +20,7 @@ public interface DayKLineDao {
 	 * @param stockID
 	 * @return
 	 */
-	List<DayKLine> queryListByStockID(int stockID);
+	List<DayKLine> queryListByStockID(long stockID);
 
 	/**
 	 * 根据 kLineID 读取对象
